Implicitly link to input.dll on Windows.
Now that we we no longer support Windows XP, we can safely assume
that input.dll is always available and it's OK to implicitly link
to that rather than relying on LoadLibrary and GetProcAddress.

One tricky part is that Windows SDK does not provide the import
library for input.dll.  To work around this, we will create our own
import library as a part of build steps, by following this document.
http://support.microsoft.com/kb/131313/en-us

BUG=none
TEST=compile